rename libsdl to libsdl2
