+++ packages/l/libsdl2/patches/2.30.0/fix_mingw.patch
